Update synapsesqlpoolexecrequests.md
There have been multiple customers in the past few weeks that have been looking for a solution to retrieve the full query text from Log Analytics.  Though it is typically understood that omission from documentation means it's not supported, customers are inquiring when it will be.  After discussions with Synapse PMs on this topic, inclusion of Command2 on this table has been rejected for performance/reliability reasons (it's too large and will cause problems); it would be preferable to provide a disclaimer statement here noting that Command2 has intentionally been left out.