Misleading description of Level column in AppServiceConsoleLogs.
For the App Service, standard output and standard error output can be obtained from AppServiceConsoleLogs. These logs are output as Standard Output = Informational level and Standard Error Output = Error level. For standard output, all output is at the Informational level, regardless of the log level at which it is output.

In the Azure Monitor reference page, the description of the Level column for AppServiceConsoleLogs is misleading, we think, because it is perceived as referring to the log level of standard output. In fact, AppServiceConsoleLogs Level is supposed to have only two types of output: standard output or standard error output.

Regards,